+++ src/Command/DeleteAllCommand.php
@@ -42,6 +42,7 @@ class DeleteAllCommand extends Command
      * @param \Cake\Console\Arguments $args The command arguments
      * @param \Cake\Console\ConsoleIo $io The console io
      * @return void
+     * @throws \ErrorException|\ReflectionException
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupShell#delete_all
      */
     public function execute(Arguments $args, ConsoleIo $io): void