+++ src/Command/ExportCommand.php
@@ -18,8 +18,6 @@ namespace DatabaseBackup\Command;
 use Cake\Console\Arguments;
 use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
-use DatabaseBackup\Command\RotateCommand;
-use DatabaseBackup\Command\SendCommand;
 use DatabaseBackup\Console\Command;
 use DatabaseBackup\Utility\BackupExport;
 use Exception;
@@ -70,8 +68,9 @@ class ExportCommand extends Command
      * @param \Cake\Console\Arguments $args The command arguments
      * @param \Cake\Console\ConsoleIo $io The console io
      * @return void
-     * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupShell#export
      * @throws \Cake\Console\Exception\StopException
+     * @throws \ReflectionException
+     * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupShell#export
      */
     public function execute(Arguments $args, ConsoleIo $io): void
     {