+++ src/Driver/Driver.php
@@ -24,8 +24,7 @@ use Symfony\Component\Process\Process;
 use Tools\Exceptionist;
 
 /**
- * Represents a driver containing all methods to export/import database backups
- *  according to the connection
+ * Represents a driver containing all methods to export/import database backups according to the connection
  * @method \Cake\Event\EventManager getEventManager()
  */
 abstract class Driver implements EventListenerInterface
@@ -36,7 +35,7 @@ abstract class Driver implements EventListenerInterface
     /**
      * @var \Cake\Database\Connection
      */
-    protected $connection;
+    protected Connection $connection;
 
     /**
      * Construct
@@ -46,17 +45,15 @@ abstract class Driver implements EventListenerInterface
     {
         $this->connection = $connection;
 
-        //Attachs the object to the event manager
+        //Attaches the object to the event manager
         $this->getEventManager()->on($this);
     }
 
     /**
-     * List of events this object is implementing. When the class is registered
-     *  in an event manager, each individual method will be associated with the
-     *  respective event
-     * @return array<string, string> Associative array or event key names pointing
-     *  to the function that should be called in the object when the respective
-     *  event is fired
+     * List of events this object is implementing. When the class is registered in an event manager, each individual
+     *  method will be associated with the respective event
+     * @return array<string, string> Associative array or event key names pointing to the function that should be called
+     *  in the object when the respective event is fired
      * @since 2.1.1
      */
     final public function implementedEvents(): array
@@ -84,15 +81,16 @@ abstract class Driver implements EventListenerInterface
     }
 
     /**
-     * Gets and parses executable commands from the configuration, according to
-     *  the type of requested operation (`export` or `import`) and the connection
-     *  driver.
+     * Gets and parses executable commands from the configuration, according to the type of requested operation
+     *  (`export` or `import`) and the connection driver.
      *
-     * These executables are not yet final, use instead `_getExportExecutable()`
-     *  and `_getImportExecutable()` methods to have the final executables,
-     *  including compression.
+     * These executables are not yet final, use instead `_getExportExecutable()` and `_getImportExecutable()` methods to
+     *  have the final executables, including compression.
      * @param string $type Type or the request operation (`export` or `import`)
      * @return string
+     * @throws \Tools\Exception\NotInArrayException
+     * @throws \ReflectionException
+     * @throws \ErrorException
      */
     protected function _getExecutable(string $type): string
     {
@@ -100,7 +98,7 @@ abstract class Driver implements EventListenerInterface
         $driver = strtolower($this->getDriverName());
         $replacements = [
             '{{BINARY}}' => escapeshellarg($this->getBinary(DATABASE_BACKUP_EXECUTABLES[$driver][$type])),
-            '{{AUTH_FILE}}' => isset($this->auth) ? escapeshellarg($this->auth) : '',
+            '{{AUTH_FILE}}' => method_exists($this, 'getAuthFile') && $this->getAuthFile() ? escapeshellarg($this->getAuthFile()) : '',
             '{{DB_USER}}' => $this->getConfig('username'),
             '{{DB_PASSWORD}}' => $this->getConfig('password') ? ':' . $this->getConfig('password') : '',
             '{{DB_HOST}}' => $this->getConfig('host'),
@@ -115,6 +113,9 @@ abstract class Driver implements EventListenerInterface
      * Gets the executable command to export the database, with compression if requested
      * @param string $filename Filename where you want to export the database
      * @return string
+     * @throws \Tools\Exception\NotInArrayException
+     * @throws \ReflectionException
+     * @throws \ErrorException
      */
     protected function _getExportExecutable(string $filename): string
     {
@@ -131,6 +132,9 @@ abstract class Driver implements EventListenerInterface
      * Gets the executable command to import the database, with compression if requested
      * @param string $filename Filename from which you want to import the database
      * @return string
+     * @throws \Tools\Exception\NotInArrayException
+     * @throws \ReflectionException
+     * @throws \ErrorException
      */
     protected function _getImportExecutable(string $filename): string
     {
@@ -195,8 +199,7 @@ abstract class Driver implements EventListenerInterface
     /**
      * Gets a config value or the whole configuration of the connection
      * @param string|null $key Config key or `null` to get all config values
-     * @return mixed Config value, `null` if the key doesn't exist
-     *  or all config values if no key was specified
+     * @return mixed Config value, `null` if the key doesn't exist or all config values if no key was specified
      * @since 2.3.0
      */
     final public function getConfig(?string $key = null)
@@ -210,9 +213,8 @@ abstract class Driver implements EventListenerInterface
      * Exports the database.
      *
      * When exporting, this method will trigger these events:
-     *
-     * - Backup.beforeExport: will be triggered before export
-     * - Backup.afterExport: will be triggered after export
+     *  - `Backup.beforeExport`: will be triggered before export;
+     *  - `Backup.afterExport`: will be triggered after export.
      * @param string $filename Filename where you want to export the database
      * @return bool `true` on success
      * @throws \Exception
@@ -236,12 +238,13 @@ abstract class Driver implements EventListenerInterface
      * Imports the database.
      *
      * When importing, this method will trigger these events:
-     *
-     * - Backup.beforeImport: will be triggered before import
-     * - Backup.afterImport: will be triggered after import
+     *  - `Backup.beforeImport`: will be triggered before import;
+     *  - `Backup.afterImport`: will be triggered after import.
      * @param string $filename Filename from which you want to import the database
      * @return bool true on success
-     * @throws \Exception
+     * @throws \Tools\Exception\NotInArrayException
+     * @throws \ReflectionException
+     * @throws \ErrorException
      */
     final public function import(string $filename): bool
     {