+++ tests/bootstrap.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -16,9 +17,10 @@ declare(strict_types=1);
 use Cake\Cache\Cache;
 use Cake\Core\Configure;
 use Cake\Datasource\ConnectionManager;
-use Cake\Mailer\Email;
+use Cake\Mailer\Mailer;
 use Cake\Mailer\TransportFactory;
 use Cake\TestSuite\TestEmailTransport;
+use DatabaseBackup\Utility\BackupExport;
 use DatabaseBackup\Utility\BackupManager;
 
 date_default_timezone_set('UTC');
@@ -83,7 +85,7 @@ Cache::setConfig([
 ]);
 
 TransportFactory::setConfig('debug', ['className' => TestEmailTransport::class]);
-Email::setConfig('default', ['transport' => 'debug']);
+Mailer::setConfig('default', ['transport' => 'debug']);
 
 if (!getenv('db_dsn')) {
     putenv('db_dsn=mysql://travis@localhost/test');
@@ -106,3 +108,36 @@ Configure::write('pluginsToLoad', ['DatabaseBackup']);
 
 require_once ROOT . 'config' . DS . 'bootstrap.php';
 echo 'Running tests for `' . BackupManager::getDriverName() . '` driver ' . PHP_EOL;
+
+if (!function_exists('createBackup')) {
+    /**
+     * Global function to create a backup file
+     * @param string $filename Filename
+     * @return string
+     * @throws \Tools\Exception\NotWritableException|\ErrorException
+     */
+    function createBackup(string $filename = 'backup.sql'): string
+    {
+        return (new BackupExport())->filename($filename)->export();
+    }
+}
+
+if (!function_exists('createSomeBackups')) {
+    /**
+     * Global function to create some backup files
+     * @return array
+     * @throws \Tools\Exception\NotWritableException|\ErrorException
+     */
+    function createSomeBackups(): array
+    {
+        $timestamp = time();
+
+        foreach (['sql.gz', 'sql.bz2', 'sql'] as $extension) {
+            $file = createBackup('backup_test_' . $timestamp . '.' . $extension);
+            touch($file, $timestamp--);
+            $files[] = $file;
+        }
+
+        return array_reverse($files);
+    }
+}